refactor to split out layout for data flow
